reorganize connection info section content, migrate danger info section
